Fix additional warnings
- Deprecation
- Unchecked (in tests)
- Eval CLI static access

+++ lenskit-data-structures/src/test/java/org/grouplens/lenskit/collections/RepeatedListTest.java
@@ -62,7 +62,7 @@ public class RepeatedListTest {
     public void testSingle() {
         RepeatedList<Integer> rl = new RepeatedList<Integer>(7, 1);
         assertThat(rl.size(), equalTo(1));
-        assertThat(((Integer)rl.get(0)).intValue(), equalTo(7));
+        assertThat(rl.get(0).intValue(), equalTo(7));
         
         try {
             rl.get(1);
